So here's what we have done today. We started writing issues for the collision detection algorithm and quickly ran into some problems with the algorithm. Don't forget that the algorithm we're using keeps track of not only future collisions, but also past collisions. As a result of this fact we needed to figure out a solution to it. We had started to write tests, but ran into those fatal errors. We also ran into the problem of multiple planes coming into the PA and found a solution to that as well. (Read Ben's email) We finally started to stub out our methods in the actual files to start tests from them. We wrote a few issues on a google doc, which we will send to John and Jesse to add to and/or start writing unit tests if they want.
John- Jesse -> If you want to start writing unit tests for those issues up on the google doc, you're more then welcome to, but make a note saying you're working on it so that no one works on the same section. We decided that the collision detection method just returns true or false based on whether there is a collision or not. The sorting of the collisions will be done in the plane manager. I am sure you will have questions, so let us know. Since we were able to sort out the algorithm even better today with some tests, tomorrow the issues and test should cook.